LPS-49545 move to module
