+++ modules/apps/portal-settings/portal-settings-authentication-ldap-web/src/main/resources/META-INF/resources/com.liferay.portal.settings.web/edit_ldap_server.jsp
@@ -1,4 +1,6 @@
-<%--
+<%@ page import="com.liferay.portal.ldap.DuplicateLDAPServerNameException" %>
+<%@ page import="com.liferay.portal.ldap.validator.LDAPFilterException" %>
+<%@ page import="com.liferay.portal.ldap.LDAPServerNameException" %><%--
 /**
  * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
