+++ modules/portal/portal-ldap/src/main/java/com/liferay/portal/ldap/internal/DefaultLDAPSettings.java
@@ -14,7 +14,6 @@
 
 package com.liferay.portal.ldap.internal;
 
-import com.liferay.portal.kernel.ldap.LDAPUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.log.LogUtil;
@@ -31,6 +30,7 @@ import com.liferay.portal.ldap.configuration.LDAPServerConfiguration;
 import com.liferay.portal.ldap.configuration.SystemLDAPConfiguration;
 import com.liferay.portal.ldap.exportimport.configuration.LDAPExportConfiguration;
 import com.liferay.portal.ldap.exportimport.configuration.LDAPImportConfiguration;
+import com.liferay.portal.ldap.util.LDAPUtil;
 
 import java.util.Properties;
