+++ modules/portal/portal-ldap/src/main/java/com/liferay/portal/ldap/internal/exportimport/LDAPUserImporterImpl.java
@@ -25,7 +25,6 @@ import com.liferay.portal.kernel.exception.NoSuchRoleException;
 import com.liferay.portal.kernel.exception.NoSuchUserGroupException;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.ldap.LDAPUtil;
 import com.liferay.portal.kernel.lock.LockManager;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -40,12 +39,7 @@ import com.liferay.portal.kernel.model.UserGroup;
 import com.liferay.portal.kernel.security.exportimport.UserGroupImportTransactionThreadLocal;
 import com.liferay.portal.kernel.security.exportimport.UserImporter;
 import com.liferay.portal.kernel.security.ldap.AttributesTransformer;
-import com.liferay.portal.kernel.security.ldap.LDAPGroup;
 import com.liferay.portal.kernel.security.ldap.LDAPSettings;
-import com.liferay.portal.kernel.security.ldap.LDAPToPortalConverter;
-import com.liferay.portal.kernel.security.ldap.LDAPUser;
-import com.liferay.portal.kernel.security.ldap.LDAPUserImporter;
-import com.liferay.portal.kernel.security.ldap.PortalLDAP;
 import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.RoleLocalService;
@@ -65,11 +59,17 @@ import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.ldap.ContactConverterKeys;
+import com.liferay.portal.ldap.PortalLDAP;
 import com.liferay.portal.ldap.UserConverterKeys;
 import com.liferay.portal.ldap.configuration.ConfigurationProvider;
 import com.liferay.portal.ldap.configuration.LDAPServerConfiguration;
+import com.liferay.portal.ldap.exportimport.LDAPGroup;
+import com.liferay.portal.ldap.exportimport.LDAPToPortalConverter;
+import com.liferay.portal.ldap.exportimport.LDAPUser;
+import com.liferay.portal.ldap.exportimport.LDAPUserImporter;
 import com.liferay.portal.ldap.exportimport.configuration.LDAPImportConfiguration;
 import com.liferay.portal.ldap.internal.UserImportTransactionThreadLocal;
+import com.liferay.portal.ldap.util.LDAPUtil;
 
 import java.io.Serializable;
