+++ modules/portal/portal-ldap/src/main/java/com/liferay/portal/ldap/validator/LDAPFilterException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.ldap;
+package com.liferay.portal.ldap.validator;
 
 import com.liferay.portal.kernel.exception.PortalException;
